Move razzia tests to razzia app
